main/requirements.txt
Django==6.0.1
Pillow==12.1.1
whitenoise==6.12.0
dj-database-url==3.1.2
psycopg2-binary==2.9.11
python-decouple==3.8
django-jazzmin==3.0.1
